+++ src/collective/timestamp/utils.py
@@ -41,7 +41,7 @@ def timestamp(
     service_url: str,
     hashing_algorithm: str = "sha256",
     use_failover: bool = False,
-    failover_timestamping_service_urls: list = (),
+    failover_timestamping_service_urls: list = None,
     max_retries: int = 0,
     initial_backoff_seconds: float = 0.5,
     localize_date: bool = True,
@@ -64,7 +64,7 @@ def timestamp(
     success = False
     tsr = None
     # Make a copy to avoid modifying the original list
-    failover_urls = failover_timestamping_service_urls.copy()
+    failover_urls = failover_timestamping_service_urls.copy() if failover_timestamping_service_urls else []
     while not success:
         retry_count = 0
         backoff_seconds = initial_backoff_seconds